+++ src/components/index.js
@@ -5,3 +5,4 @@ export * from './debounced-input'
 export * from './bouncing-dots'
 export * from './side-collapse'
 export * from './info-helpers'
+export * from './debounced-range-slider'
