+++ src/components/search/results/variable-view-layout/index.js
@@ -1,5 +1,2 @@
-export * from './variable-results'
-export * from './variable-results-helper'
-export * from './variable-table-by-study'
-export * from './variables-histogram'
+export * from './variable-results-new'
 export * from './variable-view-layout'
