+++ src/components/search/results/variable-view-layout/study-data-source-list/index.ts
@@ -1,0 +1,2 @@
+export * from './study-data-source-list'
+export * from './study-data-source-tag'
