+++ src/components/search/results/variable-view-layout/study-data-source-list/study-data-source-list.tsx
@@ -1,0 +1,17 @@
+import { Space } from 'antd'
+import { useVariableView } from '../variable-view-context'
+import { StudyDataSourceTag } from './study-data-source-tag'
+
+
+export const StudyDataSourcesList = () => {
+    const { orderedDataSources } = useVariableView()!
+    return (
+        <Space direction="horizontal" size={ 8 } wrap style={{ flexGrow: 1 }}>
+            {
+                orderedDataSources.map((dataSource) => (
+                    <StudyDataSourceTag key={ `variable-view-data-source-tag-${ dataSource.name }` } dataSource={ dataSource } />
+                ))
+            }
+        </Space>
+    )
+}
