+++ src/components/search/results/variable-view-layout/study-info-tooltip/index.ts
@@ -1,0 +1,1 @@
+export * from './study-info-tooltip'
