+++ src/components/search/results/variable-view-layout/variable-filter-menu/index.ts
@@ -1,0 +1,1 @@
+export * from './variable-filter-menu'
