+++ src/components/search/results/variable-view-layout/variable-list/index.ts
@@ -1,0 +1,3 @@
+export * from './study-list-item'
+export * from './variable-list-item'
+export * from './variable-list'
