+++ src/components/search/results/variable-view-layout/variable-view-histogram/index.ts
@@ -1,0 +1,1 @@
+export * from './variable-view-histogram'
